requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.2
pytest-django==4.8.0
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
httpx==0.27.2
mysqlclient==2.2.4
pyjwt==2.4.0